pyproject.toml
[tool.poetry]
name = "baseball.compuer"
version = "0.1.0"
description = ""
authors = ["David Roher <davidroher@davids-air.home>"]

[tool.poetry.dependencies]
python = "^3.11.0,<3.12"

dbt-core = "~1.7.2"
dbt-duckdb = "~1.7.0"
duckdb = "~0.9.2"

tensorflow = "~2.15.0"
tensorflow-metal = {version = "~1.1.0", platform = "darwin", markers = "platform_machine=='arm64'" }
tensorflow-macos = { version = "~2.15.0", platform = "darwin", markers = "platform_machine=='arm64'" }

numpy = "~1.26.2"
pandas = "~2.1.3"
polars = "^0.19.19"
pyarrow = "^14.0.1"
statsmodels = "^0.14.0"
recordlinkage = "^0.16.0"

[tool.poetry.dev-dependencies]
sqlfluff = "^2.3.5"
sqlfluff-templater-dbt = "^2.3.5"
ipykernel = "^6.25.2"
boto3 = "^1.33.2"
"ruamel.yaml" = "^0.18.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
